# src/main/resources/static/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Order Picker API",
    "description": "    This api is made to improve programming skills with the Java language, Spring Boot framework and MySQL Database, in addition\n    the api has unit test with Junit and Mockito.\n    To begin you must create a user with Role USER or EMPLOYEE and run the different endpoints where you can create products, orders and\n    make deliveries according to the role that chose. \n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/order-picker/auth/login": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "summary": "Login",
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid values entry"
          }
        }
      }
    },
...
    "/api/order-picker/users/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get a User by id",
        "operationId": "getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID to search",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Found",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Update a User by id",
        "operationId": "updateOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User Updated",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "Delete a User by id",
        "operationId": "deleteOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Deleted",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      }
    },
...
    "/api/order-picker/users/email/{email}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get a User by email",
        "operationId": "getByEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "User email to search",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Found",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      }
    },
    "/api/order-picker/users/dni/{dni}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get a User by dni",
        "operationId": "getByDni",
        "parameters": [
          {
            "name": "dni",
            "in": "path",
            "description": "User DNI to search",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Found",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserDTO": {
        "required": [
          "address",
          "dni",
          "email",
          "fullName",
          "password",
          "phone",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dni": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "address": {
            "maxLength": 100,
            "minLength": 5,
            "type": "string"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 7,
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "EMPLOYEE",
              "ADMIN"
            ]
          }
        }
      },
...
      "AuthenticationRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}